chore: move PostLayout to its own folder
